feat: rename and create utils in picker_ui
