+++ lua/fff/picker_ui/README.md
@@ -5,7 +5,7 @@ Originally a single monolithic ~2535-line file, it has been split into focused s
 
 ## Entry point
 
-**`init.lua`** — requires `require('fff.picker_ui')`.
+**`picker_ui.lua`** — requires `require('fff.picker_ui.picker_ui')`.
 
 This is the coordinator. It wires all submodules together and exposes the public API:
 
@@ -23,7 +23,7 @@ This is the coordinator. It wires all submodules together and exposes the public
 
 | Module | File | Purpose | Key exports via `init.lua` |
 |---|---|---|---|
-| **state_manager** | `state_manager.lua` | Single source of truth for all picker state | `M.state`, `M.clear_selections`, `M.reset_history_state` |
+| **picker_ui_state** | `picker_ui_state.lua` | Single source of truth for all picker state | `M.state`, `M.clear_selections`, `M.reset_history_state` |
 | **ui_creator** | `ui_creator.lua` | Creates buffers, windows, and keymaps for the picker UI | `M.create_ui`, `M.setup_buffers`, `M.setup_windows`, `M.setup_keymaps`, `M.focus_*_win`, `M.open_preview`, `M.close_preview` |
 | **search_manager** | `search_manager.lua` | Executes searches, manages pagination, handles query history | `M.update_results_sync`, `M.update_results`, `M.load_*_page`, `M.on_input_change`, `M.cycle_grep_modes`, `M.recall_query_from_history`, `M.cycle_forward_query` |
 | **renderer** | `renderer.lua` | Renders the file list, handles combo separator, scrollbar, and empty state | `M.render_list`, `M.scroll_to_bottom` |
@@ -41,7 +41,7 @@ local P = nil  -- parent module reference
 
 function M.init(parent_module) P = parent_module end
 
-local S = state_manager.state  -- shared state
+local S = picker_ui_state.state  -- shared state
 
 -- ... functions that use S.* and P.* ...
 
@@ -50,11 +50,11 @@ return M
 
 The starter module (`init.lua`) calls `module.init(M)` on each submodule, passing itself as the parent. This lets submodules call back into `init.lua` for cross-module coordination — for example, `navigation.lua` calls `P.render_list()` and `P.update_preview()` after moving the cursor.
 
-State is shared via a single table reference (`state_manager.state`). Every submodule writes to and reads from the same table — no message passing or event bus.
+State is shared via a single table reference (`picker_ui_state.state`). Every submodule writes to and reads from the same table — no message passing or event bus.
 
 ## Key design decisions
 
-- **`state_manager` is a pure data store** — no parent module reference, no `init()`. It owns the state table and selection helpers.
+- **`picker_ui_state` is a pure data store** — no parent module reference, no `init()`. It owns the state table and selection helpers.
 - **`init.lua` keeps cross-cutting concerns** — `update_status`, `select`, `toggle_debug`, `open` lives here because they coordinate across multiple submodules.
 - **Module purpose is reflected in the filename** — anything with `_manager` in the name manages state or lifecycle; `renderer` and `navigation` are purely behavioral.
 - **`vim.schedule` usage is intentional** — deferred calls prevent re-entrancy issues during buffer mutations and window teardown.
