+++ lua/fff/picker_ui/picker_ui_state.lua
@@ -209,90 +209,4 @@ function M.toggle_selection()
   return was_selected
 end
 
--- Send selected files/matches to quickfix list
-function M.send_to_quickfix()
-  if not M.state.active then return end
-
-  local qf_list = {}
-
-  if M.state.mode == 'grep' then
-    -- Grep mode: per-occurrence entries with exact locations
-    local has_selections = next(M.state.selected_items) ~= nil
-
-    if has_selections then
-      -- Use explicitly selected items (survives page changes)
-      for _, item in pairs(M.state.selected_items) do
-        local abs = canonicalize_fff_path(item.relative_path)
-        if abs then
-          table.insert(qf_list, {
-            filename = abs,
-            lnum = item.line_number or 1,
-            col = (item.col or 0) + 1,
-            text = item.line_content or vim.fn.fnamemodify(abs, ':.'),
-          })
-        end
-      end
-    else
-      -- No selections: run an exhaustive search to get all matches
-      local grep = require('fff.grep')
-      local exhaustive_config = vim.tbl_extend('force', M.state.grep_config or {}, { max_matches_per_file = 0 })
-      local exhaustive = grep.search(M.state.query, 0, 10000, exhaustive_config, M.state.grep_mode)
-      local all_items = exhaustive and exhaustive.items or {}
-
-      if #all_items == 0 then
-        vim.notify('No matches to send to quickfix', vim.log.levels.WARN)
-        return
-      end
-
-      for _, item in ipairs(all_items) do
-        local abs = canonicalize_fff_path(item.relative_path)
-        if abs then
-          table.insert(qf_list, {
-            filename = abs,
-            lnum = item.line_number or 1,
-            col = (item.col or 0) + 1,
-            text = item.line_content or vim.fn.fnamemodify(abs, ':.'),
-          })
-        end
-      end
-    end
-  else
-    -- Normal file mode: per-file entries at line 1
-    local paths = {}
-
-    -- Collect from explicit selections, or fall back to all visible items.
-    if next(M.state.selected_files) then
-      for relative_path, _ in pairs(M.state.selected_files) do
-        table.insert(paths, canonicalize_fff_path(relative_path))
-      end
-    else
-      for _, item in ipairs(M.state.filtered_items) do
-        local abs = canonicalize_fff_path(item.relative_path)
-        if abs then table.insert(paths, abs) end
-      end
-    end
-
-    if #paths == 0 then
-      vim.notify('No files to send to quickfix', vim.log.levels.WARN)
-      return
-    end
-
-    for _, path in ipairs(paths) do
-      table.insert(qf_list, {
-        filename = path,
-        lnum = 1,
-        col = 1,
-        text = vim.fn.fnamemodify(path, ':.'),
-      })
-    end
-  end
-
-  local is_grep = M.state.mode == 'grep'
-
-  -- Clear selections after sending to quickfix
-  M.clear_selections()
-
-  return qf_list, is_grep
-end
-
 return M
